Address bug in remove_principal_components & fix type hints

+++ fse/models/usif.py
@@ -22,7 +22,7 @@ class uSIF(Average):
         model: BaseKeyedVectors,
         length: int = None,
         components: int = 5,
-        cache_size_gb: int = 1.0,
+        cache_size_gb: float = 1.0,
         sv_mapfile_path: str = None,
         wv_mapfile_path: str = None,
         workers: int = 1,
